UI test cleanup: Extract iter_nth tests
